### 2023-06-11 #5540 @editorialbot
[PRE REVIEW]: ER-Evaluation: End-to-End Evaluation of Entity Resolution Systems
**Submitting author:** @OlivierBinette (<a href="http://orcid.org/0000-0001-6009-5206">Olivier Binette</a>)
**Repository:** https://github.com/Valires/er-evaluation
**Branch with paper.md** (empty if default branch):
**Version:** 2.1.0
