### 2023-06-30 #5540 @osorensen
👋 @robinL, @tedenamorado, @RossKen, @ThomasHepworth, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
