### 2023-07-01 #5540 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1126/sciadv.abi8021 is OK
> - 10.14778/3137628.3137642 is OK
> - 10.1145/3418896 is OK
>
> MISSING DOIs
>
> - 10.1080/00031305.2023.2191664 may be a valid DOI for title: Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org
>
> INVALID DOIs
>
> - None
> ```

@OlivierBinette, while we're waiting for the review to start, could you please check the missing DOI, and update your BibTeX source file if the suggested DOI is correct?
