### 2023-07-04 #5619 @editorialbot
[REVIEW]: ER-Evaluation: End-to-End Evaluation of Entity Resolution Systems
**Submitting author:** @OlivierBinette (<a href="http://orcid.org/0000-0001-6009-5206">Olivier Binette</a>)
**Repository:** https://github.com/Valires/er-evaluation
**Branch with paper.md** (empty if default branch):
**Version:** Version 2.2.1
