### 2023-07-20 #5619 @Nikoletos-K
I have completed my review. ✅

I would like to make some comments/suggestions for @OlivierBinette:
- __Automated tests:__ You do have a test, but I believe that more tests and greater code coverage would be preferable. For automation, look at codecov and GitHub actions.
- __Example usage:__ I see notebooks with experiments using datasets like PatentsView, however it would be wonderful to have your software test some ER frameworks, in a notebook for example. [pyJedAI](https://github.com/AI-team-UoA/pyJedAI) would be ideal for this!✌️
