### 2023-08-17 #5619 @ThomasHepworth
Heya and apologies for the delays. See above for my completed review.

I've run through all of the documentation and tried out the package on some synthetic data that we (the Ministry of Justice) have created. Everything works as expected and it produces some interesting insights.

My wider team are also all aware of the package and we're excited to dig into it and see how we can go about making use of it internally.

**Some minor comments for @OlivierBinette:**
I'd just like to echo Nikoletos's points, particularly around testing - https://github.com/openjournals/joss-reviews/issues/5619#issuecomment-1643791160.

I'd also like to add that it might make sense to lint some of the codebase and docs. Some of your code examples can run off the page in places.

I'll also leave a few very minor comments on the repo itself. Feel free to ignore and close these issues if you feel differently or see them as non-issues.

Thanks again!
Tom
