+++ er_evaluation/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.0"
+__version__ = "2.1.0"
 
 import er_evaluation.data_structures
 import er_evaluation.datasets
