+++ CONTRIBUTORS.md
@@ -1,20 +1,0 @@
-## Contributors
-
-<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
-<!-- prettier-ignore-start -->
-<!-- markdownlint-disable -->
-<table>
-  <tbody>
-    <tr>
-      <td align="center" valign="top" width="14.28%"><a href="https://olivierbinette.github.io/"><img src="https://avatars.githubusercontent.com/u/784901?v=4?s=100" width="100px;" alt="Olivier Binette"/><br /><sub><b>Olivier Binette</b></sub></a><br /><a href="#code-OlivierBinette" title="Code">💻</a> <a href="#content-OlivierBinette" title="Content">🖋</a> <a href="#research-OlivierBinette" title="Research">🔬</a> <a href="#ideas-OlivierBinette" title="Ideas, Planning, & Feedback">🤔</a> <a href="#doc-OlivierBinette" title="Documentation">📖</a></td>
-      <td align="center" valign="top" width="14.28%"><a href="https://github.com/ybaek"><img src="https://avatars.githubusercontent.com/u/44536898?v=4?s=100" width="100px;" alt="Youngsoo Baek"/><br /><sub><b>Youngsoo Baek</b></sub></a><br /><a href="#research-ybaek" title="Research">🔬</a> <a href="#ideas-ybaek" title="Ideas, Planning, & Feedback">🤔</a></td>
-    </tr>
-  </tbody>
-</table>
-
-<!-- markdownlint-restore -->
-<!-- prettier-ignore-end -->
-
-<!-- ALL-CONTRIBUTORS-LIST:END -->
-
-This project follows the [all-contributors](https://allcontributors.org/docs/en/emoji-key) specification. Contributions of any kind are welcome.
