### 2023-10-25 d613384
update readthedocs requirements
