+++ docs/requirements.txt
@@ -1,0 +1,25 @@
+pandas
+numpy 
+scipy
+plotly
+igraph
+scikit-learn
+wheel
+watchdog
+coverage
+Sphinx==5
+twine
+testbook
+jupyter
+ipykernel
+ipython
+black
+isort
+pydata-sphinx-theme
+sphinx_automodapi
+sphinx_design
+sphinx-mdinclude
+nbsphinx
+pandoc
+myst-nb
+graphviz
