### 2023-10-25 58fe5c1
update readthedocs requirements
