+++ .readthedocs.yaml
@@ -16,4 +16,3 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
-   system_packages: true
