### 2023-10-25 7db6833
update test dependencies
