+++ .readthedocs.yaml
@@ -16,3 +16,4 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
+
