### 2023-10-25 47d992c
update test dependencies
