### 2023-10-25 c5a6d3d
update test dependencies
