+++ .readthedocs.yaml
@@ -1,19 +1,0 @@
-# .readthedocs.yaml
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-version: 2
-
-build:
-  os: ubuntu-22.04
-  apt_packages:
-   - graphviz
-  tools:
-    python: "3.10"
-
-sphinx:
-   configuration: docs/conf.py
-
-python:
-   install:
-   - requirements: docs/requirements.txt
-
