+++ .github/workflows/python-package.yaml
@@ -16,11 +16,6 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install tox and any other packages
-        run: pip install tox
-      - name: Run tox
-        # Run tox using the version of Python in `PATH`
-        run: tox -e py
       - name: Doctest modules
         run: |
           pip install -e .[test]
