+++ setup.py
@@ -10,8 +10,7 @@ with open("README.rst") as readme_file:
 with open("CHANGELOG.rst") as history_file:
     history = history_file.read()
 
-requirements = (
-    [
+requirements = [
         "pandas",
         "numpy",
         "scipy",
@@ -21,8 +20,7 @@ requirements = (
         "pyarrow",
         "urllib3",
         "requests",
-    ],
-)
+    ]
 
 test_requirements = [
     "pytest>=3",
