+++ .readthedocs.yaml
@@ -16,3 +16,5 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
+   - method: pip
+     path: .
