+++ docs/api/er_evaluation.data_structures.isgraph.rst
@@ -1,6 +1,0 @@
-isgraph
-=======
-
-.. currentmodule:: er_evaluation.data_structures
-
-.. autofunction:: isgraph
