+++ docs/api/er_evaluation.cluster_sizes_distribution.rst
@@ -1,6 +1,0 @@
-cluster_sizes_distribution
-==========================
-
-.. currentmodule:: er_evaluation
-
-.. autofunction:: cluster_sizes_distribution
