+++ docs/api/er_evaluation.utils.load_module_parquet.rst
@@ -1,6 +1,0 @@
-load_module_parquet
-===================
-
-.. currentmodule:: er_evaluation.utils
-
-.. autofunction:: load_module_parquet
