+++ docs/api/er_evaluation.error_metrics_from_table.rst
@@ -1,6 +1,0 @@
-error_metrics_from_table
-========================
-
-.. currentmodule:: er_evaluation
-
-.. autofunction:: error_metrics_from_table
