+++ docs/api/er_evaluation.estimators.cluster_f_estimator.rst
@@ -1,6 +1,0 @@
-cluster_f_estimator
-===================
-
-.. currentmodule:: er_evaluation.estimators
-
-.. autofunction:: cluster_f_estimator
