+++ docs/api/er_evaluation.fit_dt_regressor.rst
@@ -1,6 +1,0 @@
-fit_dt_regressor
-================
-
-.. currentmodule:: er_evaluation
-
-.. autofunction:: fit_dt_regressor
