+++ docs/api/er_evaluation.data_structures.ispairs.rst
@@ -1,6 +1,0 @@
-ispairs
-=======
-
-.. currentmodule:: er_evaluation.data_structures
-
-.. autofunction:: ispairs
