+++ docs/api/er_evaluation.data_structures.isclusters.rst
@@ -1,6 +1,0 @@
-isclusters
-==========
-
-.. currentmodule:: er_evaluation.data_structures
-
-.. autofunction:: isclusters
