+++ docs/api/er_evaluation.data_structures.ismembership.rst
@@ -1,6 +1,0 @@
-ismembership
-============
-
-.. currentmodule:: er_evaluation.data_structures
-
-.. autofunction:: ismembership
