+++ docs/api/er_evaluation.summary.cluster_hill_number.rst
@@ -1,6 +1,0 @@
-cluster_hill_number
-===================
-
-.. currentmodule:: er_evaluation.summary
-
-.. autofunction:: cluster_hill_number
