+++ docs/api/er_evaluation.plots.plot_metrics.rst
@@ -1,6 +1,0 @@
-plot_metrics
-============
-
-.. currentmodule:: er_evaluation.plots
-
-.. autofunction:: plot_metrics
