+++ docs/00-overview.rst
@@ -8,13 +8,13 @@ ER-Evaluation is a Python 3.7+ package designed for evaluating the performance o
 
 Central to the package are **principled statistical estimators** for performance metrics and summary statistics, taking into account sampling processes and biases. Without those, performance metrics are generally over-optimistic and not representative of real-world performance, which can lead to performance rank reversals and poor system design. See `our blog post <https://www.valires.com/post/common-pitfalls-to-avoid-when-estimating-er-performance-metrics>`_ on the topic for more information.
 
-Our evaluation tools employ an **entity-centric** approach, using disambiguated entity clusters as the starting point of analysis. To use the package, you need:
+Our evaluation tools employ an **entity-centric** approach, using a sample of fully-resolved entities as the starting point of analysis. To use the package, you need:
 
-1. **Prediction(s)**: Predicted entity clusters for a set of records or entity mentions, usually the main output of an ER system.
+1. **Prediction(s)**: Predicted clusters for a set of records or entity mentions, usually the main output of an ER system.
     - You can a single clustering or you can compare multiple clustering results.
-2. **Reference/benchmark data**: A sample of resolved entities, usually from a trusted benchmark dataset or from data labeling.
-    - Note that each entity/cluster in the reference dataset should be *complete*, i.e. it should be fully resolve and not exclude any record or mention.
-    - In our experience, benchmark datasets including 200 to 400 disambiguated entities are generally sufficient.
+2. **Reference/benchmark data**: A sample of fully-resolved entities, usually from a trusted benchmark dataset or from data labeling.
+    - Note that each entity/cluster in the reference dataset should be *complete*, i.e. it should be fully resolved and it should not exclude any record or mention.
+    - In our experience, benchmark datasets including 200 to 400 resolved entities are generally sufficient.
 
 .. admonition:: Example
 
@@ -27,6 +27,23 @@ Our evaluation tools employ an **entity-centric** approach, using disambiguated 
     - Ensure reference data is representative of the entire population or apply sampling weights for accurate performance metrics.
 
 With predictions and reference data available, you can install the package and begin evaluation.
+
+---------------------------
+Terminology and Definitions
+---------------------------
+
+Throughout this user guide and the documentation of the package, we use the following terminology:
+
+- An **entity** is a real-world object, e.g. a person, a company, or a product. 
+- A **record** is a set of attributes describing an entity, e.g. a customer record in a CRM system. The term **record** is used interchangeably with the term **entity mention**.
+- A **database** or **file** is a collection of records/mentions.
+- **Entity resolution,** also called **disambiguation,** **deduplication,** or **record linkage** aims to create a clustering of records/mentions according to the entity which they represent.
+- The output of an entity resolution system is a **predicted clustering,** i.e. an attempt at correctly clustering records/mentions according to the entity to which they refer. There may be errors in the predicted clustering, e.g. records/mentions may be incorrectly clustered together or split into multiple clusters.
+- A **reference** dataset, or a set of **ground truth** clusters, is a clustering of mentions/records that is assumed to be correct.
+
+For more information on entity resolution, we refer the reader to [Binette & Steorts (2022)](https://www.science.org/doi/10.1126/sciadv.abi8021) and [Christophides et al. (2019)](https://arxiv.org/abs/1905.06397).
+
+[Splink](https://github.com/moj-analytical-services/splink) is state-of-the-art large-scale entity resolution software.
 
 ------------
 Installation
