+++ README.rst
@@ -9,9 +9,12 @@
         :target: https://er-evaluation.readthedocs.io/en/latest/?version=latest
         :alt: Documentation Status
 
+.. raw:: html
+
+   <div style="padding-top:20px;"></div>
 
 🔍 ER-Evaluation: An End-to-End Evaluation Framework for Entity Resolution Systems
-==================================================================================
+===================================================================================
 
 `ER-Evaluation <https://er-evaluation.readthedocs.io/en/latest>`_ is a Python package for the evaluation of entity resolution (ER) systems.
 
@@ -25,7 +28,7 @@ It provides an **entity-centric** approach to evaluation. Given a sample of reso
 
 For more information on how to resolve a sample of entities for evaluation and model training, please refer to our `data labeling guide <https://er-evaluation.readthedocs.io/en/latest/06-data-labeling.html>`_.
 
-💻 Installation
+Installation
 ---------------
 
 Install the released version from PyPI using:
@@ -40,12 +43,12 @@ Or install the development version using:
     pip install git+https://github.com/Valires/er-evaluation.git
 
 
-📖 Documentation
+Documentation
 ----------------
 
 Please refer to the documentation website `er-evaluation.readthedocs.io <https://er-evaluation.readthedocs.io/en/latest>`_.
 
-🖼️ Usage Examples
+Usage Examples
 -----------------
 
 Please refer to the `User Guide <https://er-evaluation.readthedocs.io/en/latest/userguide.html>`_ or our `Visualization Examples <https://er-evaluation.readthedocs.io/en/latest/visualizations.html>`_ for a complete usage guide.
@@ -102,12 +105,12 @@ In summary, here's how you might use the package.
 .. image:: media/plot_decisiontree.png
    :width: 400
 
-💭 Development Philosophy
+Development Philosophy
 -------------------------
 
 **ER-Evaluation** is designed to be a unified source of evaluation tools for entity resolution systems, adhering to the Unix philosophy of simplicity, modularity, and composability. The package contains Python functions that take standard data structures such as pandas Series and DataFrames as input, making it easy to integrate into existing workflows. By importing the necessary functions and calling them on your data, you can easily use ER-Evaluation to evaluate your entity resolution system without worrying about custom data structures or complex architectures.
 
-📜 Citation
+Citation
 -----------
 
 Please acknowledge the publications below if you use ER-Evaluation:
@@ -116,7 +119,7 @@ Please acknowledge the publications below if you use ER-Evaluation:
 - Binette, Olivier, Sokhna A York, Emma Hickerson, Youngsoo Baek, Sarvo Madhavan, Christina Jones. (2022). Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org. arXiv e-prints: `arxiv:2210.01230 <https://arxiv.org/abs/2210.01230>`_
 - Upcoming: "An End-to-End Framework for the Evaluation of Entity Resolution Systems With Application to Inventor Name Disambiguation"
 
-📝 Public License
+Public License
 -----------------
 
 * `GNU Affero General Public License v3 <https://www.gnu.org/licenses/agpl-3.0.en.html>`_
