+++ AUTHORS.rst
@@ -11,3 +11,16 @@ Contributors
 ------------
 
 * https://github.com/ybaek - Research, Ideas, & Feedback
+
+How to Cite
+-----------
+
+**Package:**
+
+- Binette, Olivier. (2022). ER-Evaluation: An End-to-End Evaluation Framework for Entity Resolution Systems. Available online at `github.com/Valires/ER-Evaluation <https://github.com/Valires/ER-Evaluation>`_
+
+**Research Papers:**
+
+- `Binette, Olivier, Sokhna A. York, Emma Hickerson, Youngsoo Baek, Sarvo Madhavan, and Christina Jones. "Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView. org." The American Statistician (2023): 1-11. <https://arxiv.org/abs/2210.01230>`_
+- `Binette, Olivier, Sarvo Madhavan, Jack Butler, Beth Anne Card, Emily Melluso, and Christina Jones. "PatentsView-Evaluation: Evaluation Datasets and Tools to Advance Research on Inventor Name Disambiguation." arXiv preprint arXiv:2301.03591 (2023). <https://arxiv.org/pdf/2301.03591>`_
+- Upcoming: "An End-to-End Framework for the Evaluation of Entity Resolution Systems With Application to Inventor Name Disambiguation"
