+++ .github/workflows/python-package.yaml
@@ -18,7 +18,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Doctest modules
         run: |
-          pip install .[test]
+          pip install -e .[test]
           pytest --doctest-modules --ignore-glob=examples/*
         
       - name: Pytest
