### 2023-10-26 a4b7361
add dataset tests; update workflow for distributiont tests
