+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+2.2.0 (October 26, 2023)
+----------------------
+* Streamline package structure
+* Additional tests
+* Improved documentation
+
 2.1.0 (June 02, 2023)
 ----------------------
 
