+++ MANIFEST.in
@@ -1,5 +1,4 @@
 include AUTHORS.rst
-include CONTRIBUTING.rst
 include CHANGELOG.rst
 include LICENSE
 include README.rst
