### 2023-10-26 3beeac6
Sync version 2.2.0 (#16)
* update readthedocs

* remove api docs from repo

* Improvements to docs

* Improve documentation website

* update credits page

* minor updates

* add more tests

* update test workflow

* update test workflow

* add dataset tests; update workflow for distributiont tests

* prepare release of version 2.2.0
