+++ er_evaluation/estimators/__init__.py
@@ -10,9 +10,9 @@ Representative performance estimators are necessary for accurate evaluation of e
 1. It is typically infeasible to manually label enough data to cover an entire population of interest.
 2. Naively computing performance metrics on benchmark datasets leads to highly biased and over-optimistic results that are not representative of real-world performance. This is due to the non-linear scaling of entity resolution: while it might be easy to disambiguate a small benchmark dataset, the complexity of the problem grows quadratically in the dataset size.
 
-Using small, weighted samples of ground truth data in conjunction with representative statistical estimators addresses these two issues, by providing an efficient and accurate way of evaluating the performance of entity resolution algorithms on large datasets while also taking into account the sampling processes and biases.
+To address these issues, we use performance estimators based on a weighted sample of ground truth clusters. This provides an efficient and accurate way of evaluating the performance of entity resolution algorithms on large datasets, while also taking into account sampling processes and biases.
 
-Note: In order to obtain representative performance estimators, the set of predicted clusters given as an argument to estimator functions should cover the entire population of interest. Typically, this set of predicted clusters will be much larger than the set of sampled clusters.
+**Note:** In order to obtain representative performance estimators, the set of predicted clusters given as an argument to estimator functions should cover the entire population of interest. Typically, this set of predicted clusters will be much larger than the set of sampled clusters.
 """
 from er_evaluation.estimators._estimators import (
     b_cubed_precision_estimator,
