+++ er_evaluation/metrics/__init__.py
@@ -3,16 +3,20 @@
 Evaluation Metrics (Precision, Recall, B-cubed, etc)
 ====================================================
 
-The **metrics** module provides a set of functions to compute performance evaluation metrics for entity resolution algorithms. These metrics are computed on the intersection of a predicted disambiguation and a benchmark dataset. The available metrics are:
+The **metrics** module provides a set of functions to compute common comparison metrics for entity resolution algorithms. These compare a **predicted** clustering to a **reference** clustering. The available metrics are:
 
-- **pairwise precision**: the proportion of correctly predicted links among all predicted links.
-- **pairwise recall**: the proportion of correctly predicted links among all true links.
-- **cluster precision**: the proportion of correctly predicted clusters among all predicted clusters.
-- **cluster recall**: the proportion of correctly predicted clusters among all true clusters.
+- **pairwise precision**: the proportion of predicted links that also appear in the reference clustering.
+- **pairwise recall**: the proportion of links in the reference clustering that also appear in the predicted clustering.
+- **cluster precision**: the proportion of predicted clusters that are present in the reference clustering.
+- **cluster recall**: the proportion of reference clusters that are also present in the predicted clustering.
 - **F-scores**: harmonic mean of precision and recall
 - **B-cubed metrics**: These are defined as a weighted average of mention-level measures of precision and recall.
 
-Note that these metrics are not representative of real-world performance, they only describe the performance for disambiguating the given benchmark dataset. When using these metrics, it is important to use a benchmark dataset that is representative of the population of entities being studied, otherwise, the results may not generalize.
+**Notes:** 
+
+- When computing these metrics, **all functions in this module first compute the inner join** of the reference and predicted clusterings. As such, only records that appear in both clusterings are accounted for by these metrics. All other records are dropped.
+- Records with NA cluster identifier in the reference or predicted clusterings are dropped.
+- The metrics in this module do not provide representative performance estimates. They are only useful for comparing two clusterings, such as a. For representative performance estimates, see the :mod:`er_evaluation.estimators` module.
 """
 from er_evaluation.metrics._metrics import (
     adjusted_rand_score,
