+++ docs/api/er_evaluation.error_analysis.error_indicator.rst
@@ -1,6 +1,0 @@
-error_indicator
-===============
-
-.. currentmodule:: er_evaluation.error_analysis
-
-.. autofunction:: error_indicator

[…]
