### 2023-10-26 #5619 @OlivierBinette
Hi @osorensen, I have addressed the reviewers comments. Specifically:

- I have doctests for nearly all functions, and pytest doctests are incorporated in a Github Actions testing workflow.
- I have added pytest unit tests to cover a broader range of behavior in foundational elements of the package, namely for data structures, error metrics, and data loading used throughout. Pytest of the distribution version is in my Github Actions workflow.
- I formated code using Black (line-length=120), and linted with ruff.
- I cleaned up many files and streamlined the package.
- I made minor improvements to the documentation throughout.

I am planning to add more examples, tutorials, and tests as I finish the full methodology paper, but I think these changes might already satisfy the reviewer comments.
