### 2023-10-26 8836ad0
fix rst for PyPI
