### 2023-10-26 4d46296
fix rst for PyPI
