### 2023-10-27 #5619 @osorensen
Thanks @OlivierBinette!

@Nikoletos-K, you have one unchecked item in your [review checklist](https://github.com/openjournals/joss-reviews/issues/5619#issuecomment-1643646645). Could you please consider if the latest changes made by @OlivierBinette satisfies this?
