### 2023-10-27 1bee4cf
PyPI fixes (#17)
* fix rst for PyPI

* fix rst for PyPI
