### 2023-11-01 #5619 @osorensen
@Nikoletos-K, a gentle reminder about your JOSS review. Could you please check if the latest changes satisfy the last unranked item in your [review checklist](https://github.com/openjournals/joss-reviews/issues/5619#issuecomment-1643646645)?
