### 2023-11-01 #5619 @Nikoletos-K
Hi, yes I'll proceed to re-evaluation and I'll inform you by the end of the week. Is that ok?
