### 2023-11-06 Valires/er-evaluation#18 @osorensen
Editorial comment on JOSS submission
We see

```

OK DOIs

- 10.1126/sciadv.abi8021 is OK
- 10.14778/3137628.3137642 is OK
- 10.1145/3418896 is OK

MISSING DOIs

- 10.1080/00031305.2023.2191664 may be a valid DOI for title: Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org

INVALID DOIs

- None
```

Could you please check, and add the DOI if it is correct?

https://github.com/openjournals/joss-reviews/issues/5619
