+++ paper.bib
@@ -11,56 +11,64 @@
 }
 
 @article{binette2022b,
-    author = {Binette, Olivier and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
-    journal = {arXiv e-prints},
+    author = {Olivier Binette, Sokhna A York, Emma Hickerson, Youngsoo Baek, Sarvo Madhavan and Christina Jones},
     title = {Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org},
-    year = {2022}
+    journal = {The American Statistician},
+    volume = {77},
+    number = {4},
+    pages = {370-380},
+    year = {2023},
+    publisher = {Taylor & Francis},
+    doi = {10.1080/00031305.2023.2191664}
 }
 
 @article{binette2022c,
     author = {Binette, Olivier and Madhavan, Sarvo and Butler, Jack and Card, Beth Anne and Melluso, Emily and Jones, Christina},
     journal = {arXiv e-prints},
     title = {{PatentsView-Evaluation}: {E}valuation Datasets and Tools to Advance Research on Inventor Name Disambiguation},
-    year = {2023}
+    year = {2023},
+    note = {arxiv:2301.03591},
+    doi = { 10.48550/arXiv.2301.03591}
 }
 
-@article{wang2022,
-    archivePrefix = {arXiv},
-    arxivId = {2205.05889},
-    author = {Wang, Tianshu and Lin, Hongyu and Fu, Cheng and Han, Xianpei and Sun, Le and Xiong, Feiyu and Chen, Hui and Lu, Minlong and Zhu, Xiuwen},
-    eprint = {2205.05889},
-    mendeley-groups = {ER Evaluation},
-    title = {Bridging the Gap Between Reality and Ideality of Entity Matching: A Revisiting and Benchmark Re-Construction},
-    note={arxiv:2205.05889},
-    year = {2022}
+@inproceedings{wang2022,
+  title     = {Bridging the Gap between Reality and Ideality of Entity Matching: A Revisting and Benchmark Re-Constrcution},
+  author    = {Wang, Tianshu and Lin, Hongyu and Fu, Cheng and Han, Xianpei and Sun, Le and Xiong, Feiyu and Chen, Hui and Lu, Minlong and Zhu, Xiuwen},
+  booktitle = {Proceedings of the Thirty-First International Joint Conference on Artificial Intelligence, {IJCAI-22}},
+  publisher = {International Joint Conferences on Artificial Intelligence Organization},
+  editor    = {Lud De Raedt},
+  pages     = {3978--3984},
+  year      = {2022},
+  month     = {7},
+  note      = {Main Track},
+  doi       = {10.24963/ijcai.2022/552},
+  url       = {https://doi.org/10.24963/ijcai.2022/552},
 }
 
 @book{christen2012,
-	address = {Berlin Heidelberg},
-	series = {Data-{Centric} {Systems} and {Applications}},
-	title = {Data Matching: Concepts and Techniques for Record Linkage, Entity Resolution, and Duplicate Detection},
-	shorttitle = {Data {Matching}},
-	language = {en},
-	publisher = {Springer-Verlag},
-	author = {Christen, Peter},
-	year = {2012},
-	pages = {270}
+author = {Christen, Peter},
+title = {Data Matching: Concepts and Techniques for Record Linkage, Entity Resolution, and Duplicate Detection},
+year = {2012},
+isbn = {3642311636},
+publisher = {Springer Publishing Company, Incorporated},
 }
 
 @article{marchant2017,
-    archivePrefix = {arXiv},
-    arxivId = {1703.00617},
[…]
