### 2023-11-07 Valires/er-evaluation#18 @OlivierBinette
@osorensen Thank you for your comments. I have updated the bibliography with DOIs.
