### 2023-11-07 #5619 @osorensen
@OlivierBinette, can you please complete the author items in the [checklist above](https://github.com/openjournals/joss-reviews/issues/5619#issuecomment-1798856823), and report here when done?
