+++ paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{binette2022b,
-    author = {Olivier Binette, Sokhna A York, Emma Hickerson, Youngsoo Baek, Sarvo Madhavan and Christina Jones},
+    author = {Binette, Olivier, and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
     title = {Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org},
     journal = {The American Statistician},
     volume = {77},
@@ -31,10 +31,19 @@
     doi = { 10.48550/arXiv.2301.03591}
 }
 
+@unpublished{Binette2023,
+  author = {Binette, Olivier and Baek, Youngsoo and Melluso, Emily and Jones, Christina and Dasylva, Abel and Reiter, Jerome P},
+  title = {An End-to-End Evaluation Framework for Entity Resolution Systems With Application to Inventor Name Disambiguation},
+  note = {Manuscript in progress},
+  year = {2023}
+}
+
+
+
 @inproceedings{wang2022,
-  title     = {Bridging the Gap between Reality and Ideality of Entity Matching: A Revisting and Benchmark Re-Constrcution},
+  title     = {Bridging the Gap between Reality and Ideality of Entity Matching: A Revisting and Benchmark Re-Construction},
   author    = {Wang, Tianshu and Lin, Hongyu and Fu, Cheng and Han, Xianpei and Sun, Le and Xiong, Feiyu and Chen, Hui and Lu, Minlong and Zhu, Xiuwen},
-  booktitle = {Proceedings of the Thirty-First International Joint Conference on Artificial Intelligence, {IJCAI-22}},
+  booktitle = {{P}roceedings of the {T}hirty-{F}irst {I}nternational {J}oint {C}onference on {A}rtificial {I}ntelligence, {IJCAI-22}},
   publisher = {International Joint Conferences on Artificial Intelligence Organization},
   editor    = {Lud De Raedt},
   pages     = {3978--3984},
@@ -47,9 +56,8 @@
 
 @book{christen2012,
 author = {Christen, Peter},
-title = {Data Matching: Concepts and Techniques for Record Linkage, Entity Resolution, and Duplicate Detection},
+title = {{D}ata {M}atching: {C}oncepts and {T}echniques for {R}ecord {L}inkage, {E}ntity {R}esolution, and {D}uplicate {D}etection},
 year = {2012},
-isbn = {3642311636},
 publisher = {Springer Publishing Company, Incorporated},
 }
 
@@ -72,9 +80,8 @@ publisher = {Springer Publishing Company, Incorporated},
 }
 
 @book{papadakis2021,
-  title={The Four Generations of Entity Resolution},
+  title={{T}he {F}our {G}enerations of {E}ntity {R}esolution},
   author={Papadakis, George and Ioannou, Ekaterini and Thanos, Emanouil and Palpanas, Themis},
-  pages={1--170},
   year={2021},
   publisher={Morgan \& Claypool Publishers},
 }
@@ -84,10 +91,9 @@ publisher = {Springer Publishing Company, Incorporated},
     doi = {10.1145/3418896},
     issn = {15577341},
     journal = {ACM Computing Surveys},
-    keywords = {Entity blocking and matching,batch and incremental entity resolution workflows,block processing,crowdsourcing,deep learning,strongly and nearly similar entities},
     number = {6},
     title = {An Overview of End-to-End Entity Resolution for Big Data},
     volume = {53},
     year = {2021},
-    pages={1--2}
+    pages={1--42}
 }
