+++ paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{binette2022b,
-    author = {Binette, Olivier, and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
+    author = {Binette, Olivier and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
     title = {Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org},
     journal = {The American Statistician},
     volume = {77},
