+++ paper.bib
@@ -1,99 +1,76 @@
 @article{binette2022a,
-    author = {Binette, Olivier and Steorts, Rebecca C},
-    doi = {10.1126/sciadv.abi8021},
-    journal = {Science Advances},
-    number = {12},
-    pages = {eabi8021},
-    title = {{(Almost) all of entity resolution}},
-    url = {https://www.science.org/doi/abs/10.1126/sciadv.abi8021},
-    volume = {8},
-    year = {2022}
+	title        = {{(Almost) all of entity resolution}},
+	author       = {Binette, Olivier and Steorts, Rebecca C},
+	year         = 2022,
+	journal      = {Science Advances},
+	volume       = 8,
+	number       = 12,
+	pages        = {eabi8021},
+	doi          = {10.1126/sciadv.abi8021},
 }
-
 @article{binette2022b,
-    author = {Binette, Olivier and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
-    title = {Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org},
-    journal = {The American Statistician},
-    volume = {77},
-    number = {4},
-    pages = {370-380},
-    year = {2023},
-    publisher = {Taylor & Francis},
-    doi = {10.1080/00031305.2023.2191664}
+	title        = {Estimating the Performance of Entity Resolution Algorithms: Lessons Learned Through PatentsView.org},
+	author       = {Binette, Olivier and York, Sokhna A and Hickerson, Emma and Baek, Youngsoo and Madhavan, Sarvo and Jones, Christina},
+	year         = 2023,
+	journal      = {The American Statistician},
+	publisher    = {Taylor & Francis},
+	volume       = 77,
+	number       = 4,
+	pages        = {370--380},
+	doi          = {10.1080/00031305.2023.2191664}
 }
-
 @article{binette2022c,
-    author = {Binette, Olivier and Madhavan, Sarvo and Butler, Jack and Card, Beth Anne and Melluso, Emily and Jones, Christina},
-    journal = {arXiv e-prints},
-    title = {{PatentsView-Evaluation}: {E}valuation Datasets and Tools to Advance Research on Inventor Name Disambiguation},
-    year = {2023},
-    note = {arxiv:2301.03591},
-    doi = { 10.48550/arXiv.2301.03591}
+	title        = {{PatentsView-Evaluation}: {E}valuation Datasets and Tools to Advance Research on Inventor Name Disambiguation},
+	author       = {Binette, Olivier and Madhavan, Sarvo and Butler, Jack and Card, Beth Anne and Melluso, Emily and Jones, Christina},
+	year         = 2023,
+	journal      = {arXiv e-prints},
+	doi          = {10.48550/arXiv.2301.03591},
+	note         = {arxiv:2301.03591}
 }
-
 @unpublished{Binette2023,
-  author = {Binette, Olivier and Baek, Youngsoo and Melluso, Emily and Jones, Christina and Dasylva, Abel and Reiter, Jerome P},
-  title = {An End-to-End Evaluation Framework for Entity Resolution Systems With Application to Inventor Name Disambiguation},
-  note = {Manuscript in progress},
-  year = {2023}
+	title        = {An End-to-End Evaluation Framework for Entity Resolution Systems With Application to Inventor Name Disambiguation},
+	author       = {Binette, Olivier and Baek, Youngsoo and Melluso, Emily and Jones, Christina and Dasylva, Abel and Reiter, Jerome P},
+	year         = 2023,
+	note         = {Manuscript in progress}
 }
-
-
-
 @inproceedings{wang2022,
[…]
