+++ paper.bib
@@ -27,7 +27,7 @@
 	doi          = {10.48550/arXiv.2301.03591},
 	note         = {arxiv:2301.03591}
 }
-@unpublished{Binette2023,
+@unpublished{binette2023,
 	title        = {An End-to-End Evaluation Framework for Entity Resolution Systems With Application to Inventor Name Disambiguation},
 	author       = {Binette, Olivier and Baek, Youngsoo and Melluso, Emily and Jones, Christina and Dasylva, Abel and Reiter, Jerome P},
 	year         = 2023,
