### 2023-11-08 2c00d69
Update paper (#20)
* Update paper.bib

* Update paper.md

* Update paper.bib

* Update paper.bib

tidy

* Update paper.bib

* Update paper.md
