### 2023-11-08 #5619 @OlivierBinette
@osorensen I completed the author items in the checklist.

Released version with latest changes is [version 2.2.1](https://pypi.org/project/ER-Evaluation/). DOI is [10.5281/zenodo.10086103](https://doi.org/10.5281/zenodo.10086103).

* [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. VERSION 2.2.1

* [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. DOI [10.5281/zenodo.10086103](https://doi.org/10.5281/zenodo.10086103)
